LPS-55267 Moves action to the proper place
