+++ portal-impl/src/com/liferay/portlet/sites/action/EditGroupAction.java
@@ -55,7 +55,6 @@ import com.liferay.portal.model.GroupConstants;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.model.LayoutConstants;
 import com.liferay.portal.model.LayoutSet;
-import com.liferay.portal.model.LayoutSetPrototype;
 import com.liferay.portal.model.MembershipRequest;
 import com.liferay.portal.model.MembershipRequestConstants;
 import com.liferay.portal.model.Role;
@@ -66,8 +65,6 @@ import com.liferay.portal.security.auth.RemoteAuthException;
 import com.liferay.portal.service.GroupLocalServiceUtil;
 import com.liferay.portal.service.GroupServiceUtil;
 import com.liferay.portal.service.LayoutLocalServiceUtil;
-import com.liferay.portal.service.LayoutSetLocalServiceUtil;
-import com.liferay.portal.service.LayoutSetPrototypeServiceUtil;
 import com.liferay.portal.service.LayoutSetServiceUtil;
 import com.liferay.portal.service.MembershipRequestLocalServiceUtil;
 import com.liferay.portal.service.MembershipRequestServiceUtil;
@@ -181,9 +178,6 @@ public class EditGroupAction extends PortletAction {
 			}
 			else if (cmd.equals(Constants.DELETE)) {
 				deleteGroups(actionRequest);
-			}
-			else if (cmd.equals("reset_merge_fail_count_and_merge")) {
-				resetMergeFailCountAndMerge(actionRequest);
 			}
 
 			sendRedirect(actionRequest, actionResponse, redirect);
@@ -396,54 +390,6 @@ public class EditGroupAction extends PortletAction {
 		}
 
 		return teams;
-	}
-
-	/**
-	 * Resets the number of failed merge attempts for the site template, which
-	 * is accessed by retrieving the layout set prototype ID. Once the counter
-	 * is reset, the modified site template is merged back into its linked site,
-	 * which is accessed by retrieving the group ID and private layout set.
-	 *
-	 * <p>
-	 * If the number of failed merge attempts is not equal to zero after the
-	 * merge, an error key is submitted to {@link SessionErrors}.
-	 * </p>
-	 *
-	 * @param  actionRequest the portlet request used to retrieve parameters
-	 * @throws Exception if an exception occurred
-	 */
-	protected void resetMergeFailCountAndMerge(ActionRequest actionRequest)
-		throws Exception {
-
-		long layoutSetPrototypeId = ParamUtil.getLong(
-			actionRequest, "layoutSetPrototypeId");
-
-		LayoutSetPrototype layoutSetPrototype =
-			LayoutSetPrototypeServiceUtil.getLayoutSetPrototype(
-				layoutSetPrototypeId);
-
-		SitesUtil.setMergeFailCount(layoutSetPrototype, 0);
-
-		long groupId = ParamUtil.getLong(actionRequest, "groupId");
-		boolean privateLayoutSet = ParamUtil.getBoolean(
-			actionRequest, "privateLayoutSet");
-
-		LayoutSet layoutSet = LayoutSetLocalServiceUtil.getLayoutSet(
-			groupId, privateLayoutSet);
-
-		SitesUtil.resetPrototype(layoutSet);
-
-		Group group = GroupLocalServiceUtil.getGroup(groupId);
-
-		SitesUtil.mergeLayoutSetPrototypeLayouts(group, layoutSet);
-
-		layoutSetPrototype =
-			LayoutSetPrototypeServiceUtil.getLayoutSetPrototype(
-				layoutSetPrototypeId);
-
-		if (SitesUtil.getMergeFailCount(layoutSetPrototype) > 0) {
-			SessionErrors.add(actionRequest, "resetMergeFailCountAndMerge");
-		}
 	}
 
 	protected void updateActive(ActionRequest actionRequest, String cmd)
